Fixed speed freezing
Fixed speed freezing the best that I could. only time it freezes now is when you been running speed 2x for a while as it needs to catch up the time line label. so just give it a min your program will prob freeze for a little but depending on how long you were in speed up for.